requirements.txt
requests>=2.28.0
python-dotenv>=1.0.0
unidiff>=0.7.5
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.0.0
websockets>=12.0
